Use a single activity for both asset and mission selection
This reduces the jumping around that occurs between activities
