+++ app/src/main/java/org/canterburyairpatrol/smmclient/ui/activity/MainSelector.kt
@@ -12,7 +12,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
@@ -25,16 +24,21 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import org.canterburyairpatrol.smmclient.ConnectionSingleton
+import org.canterburyairpatrol.smmclient.MainActivity
+import org.canterburyairpatrol.smmclient.R
+import org.canterburyairpatrol.smmclient.smm.data.SMMAsset
 import org.canterburyairpatrol.smmclient.smm.data.SMMMission
 import org.canterburyairpatrol.smmclient.ui.theme.SmmclientandroidTheme
 
-class MissionSelectorActivity : ComponentActivity() {
+class MainSelectorActivity : ComponentActivity() {
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         val connectionSingleton = ConnectionSingleton.getInstance()
         val connectionDetails = connectionSingleton.getConnectionDetails()
-
 
         setContent {
             SmmclientandroidTheme {
@@ -45,17 +49,13 @@ class MissionSelectorActivity : ComponentActivity() {
                 ) {
                     Column {
                         Row {
-                            Text(connectionDetails.username + "@" + connectionDetails.serverURL)
+                            Text(connectionDetails?.username + "@" + connectionDetails?.serverURL)
                         }
-                        Row {
-                            Button(onClick = {
-                                startActivity(Intent(this@MissionSelectorActivity, AssetSelectorActivity::class.java))
-                            }) {
-                                Text("Assets")
-                            }
-                        }
-                        Row (Modifier.weight(1f).align(Alignment.Start)) {
-                            this@MissionSelectorActivity.MissionSelector()
+                        Row(
+                            Modifier
+                                .weight(1f)
+                                .align(Alignment.Start)) {
+                            this@MainSelectorActivity.mainContent()
                         }
                     }
                 }
@@ -63,7 +63,74 @@ class MissionSelectorActivity : ComponentActivity() {
         }
     }
 
-    @OptIn(ExperimentalMaterial3Api::class)
+    @Composable
+    fun mainContent()
+    {
+        var assets by remember { mutableStateOf(true) }
+
+        Column {
+            Row {
+                Button(onClick = {
+                    assets = true
+                }) {
+                    Text(stringResource(R.string.assets))
+                }
+                Button(onClick = {
+                    assets = false
+                }) {
+                    Text(stringResource(R.string.missions))
+                }
+                Button(onClick = {
+                    intent = Intent(this@MainSelectorActivity, MainActivity::class.java)
+                    intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
+
+                    startActivity(intent)
+                }) {
+                    Text (stringResource(R.string.logout))
+                }
+            }
+            Row {
+                if (assets)
+                {
+                    AssetSelector()
+                }
+                else
+                {
+                    MissionSelector()
+                }
+            }
+        }
+    }
+
+    @Composable
+    fun AssetListItem(asset: SMMAsset, modifier: Modifier) {
+        ListItem(
+            headlineContent = { Text(asset.name) },
+            modifier = modifier.fillMaxWidth()
+        )
+    }
+    @Composable
+    fun AssetSelector() {
+        var assetList by remember { mutableStateOf(listOf<SMMAsset>()) }
+
+        LaunchedEffect(assetList) {
+            val connectionSingleton = ConnectionSingleton.getInstance()
+            val api = connectionSingleton.getAPI()
+            assetList = (api!!.getAssetsMine().assets)
+        }
+
+        LazyColumn {
+            items(assetList) { asset ->
+                AssetListItem(asset = asset,
+                    modifier = Modifier.clickable {
+                        var intent = Intent(this@MainSelectorActivity, AssetActivity::class.java)
+                        intent.putExtra("assetDetails", asset)
+                        startActivity(intent)
+                    })
+            }
+        }
+    }
+
     @Composable
     fun MissionListItem(mission: SMMMission, modifier: Modifier) {
         ListItem(
@@ -78,14 +145,14 @@ class MissionSelectorActivity : ComponentActivity() {
         LaunchedEffect(missionList) {
             val connectionSingleton = ConnectionSingleton.getInstance()
             val api = connectionSingleton.getAPI()
-            missionList = (api.getMissionsActive().missions)
+            missionList = (api!!.getMissionsActive().missions)
         }
 
         LazyColumn {
             items(missionList) { mission ->
                 MissionListItem(mission = mission,
                     modifier = Modifier.clickable {
-                        var intent = Intent(this@MissionSelectorActivity, MissionActivity::class.java)
+                        var intent = Intent(this@MainSelectorActivity, MissionActivity::class.java)
                         intent.putExtra("missionDetails", mission)
                         startActivity(intent)
                     })
